+++ internal/web/public/assets/swimlanesDiagram-VR7AAH4N-BjwVVHYE.js
@@ -1,0 +1,8 @@
+import{c as r,s as e}from"./flowDiagram-HODETNUW-cKIG4zHV.js";import{_ as a}from"./mermaid.core-Bs2Y6iQt.js";import"./chunk-5VM5RSS4-_N7aJtSt.js";import"./chunk-XXDRQBXY-CY2NLeSo.js";import"./chunk-POPQ4Y6H-DPIZDqa-.js";import"./chunk-F27PBJKO-C2hlAw1N.js";import"./channel-BwMvYUfc.js";import"./index-DlPkVaet.js";var o=a(t=>`${e(t)}
+  .swimlane.cluster rect {
+    stroke: ${t.clusterBorder} !important;
+  }
+  [data-look="neo"].cluster rect {
+    filter: none;
+  }
+`,"getStyles"),s=o,n=r({defaultLayout:"swimlane",styles:s});export{n as diagram};